relocate_sdk.patch: Fix file sections overwrites
The script relocate_sdk.py contains two bugs that can cause
to install the SDK with incorrectly relocated/corrupted files:
If installing the SDK into a non-default location and if the path
to the new SDK location is longer than the path of the default SDK location,
we will overwrite some data inadvertently. This can occur when
relocating sections .ldsocache and .gccrelocprefix.

One of the two bugs (overwriting of .gccrelocprefix) was fixed by an earlier
patch. That is now not needed as this new new patch provides fixes for both
sections overwrites and some additional error checks as well.

JIRA: ZEP-75

Change-Id: If86ff9d26d3e95a6cda63791c53dbe166c223c5c
Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>