hidapi-libusb: HIDAPI library for OpenOCD
A HIDAPI library is needed by OpenOCD to support cmsis-dap.
There are two implementations of this library. Here we build
the libusb immplementation, which is based on the libusb-1.0 library.

This library needs "libusb.h" in order to build. Normally, we would
bring this file in by specifying build dependencies on libudev,
however this would bring in all kinds of other dependencies,
of which some do not support nativesdk builds. So the simplest way is
to provide the file directly.

Change-Id: I49079795d1b0dbba143274e739413ea02173005f
Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>