qemu Nios2: Fix interrupt handling
Modification of control registers STATUS or IENABLE can trigger
interrupts. The modifications were not properly handled, so
on some occassions an expected interrupt was not triggered.
This patch tracks changes to the above registers and calls the
interrupt controller to evalute the changes.
(This also required to modify the interrupt controller)

Change-Id: Id65ab181bed37e1cffbe9fb432cd16da0df1209d
Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>