template_dir: Installed SDK version identification
Once an SDK is installed, it is not easy to determine the installed
SDK version.
This patch allows an easy SDK version identification by creating
a folder "info-zephyr-sdk-<version>", i.e."info-zephyr-sdk-0.8.1".
The folder currently contains versions of all toolchains, but it is
expected to contain additional information files in the future.

ZEP-387

Change-Id: I168a961d41741f91c9f465a4703dd37de630b76c
Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>