template_dir: Setup thumb2 legacy symlinks
As of SDK 0.8.2 ARM multilibs were modified/moved into changed folder names.
In order not to break any existing Zephyr Makefiles which still expect
the old location of the libraries, we set up symlinks that point
to the new locations.

In particular, we need to ensure thumb/thumb2 exists as it is referenced
directly from Zephyr's Makefile.toolchain.zephyr. If it does not, we point
 "thumb/thumb2" to "armv7-m" (providing it exists).

Change-Id: I77e0ab73c0ae3f5f237acfcb4d1bc12e95e1a4d5
Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>