newlib: support for nano-formatted-io.
If newlib is configured with --enable-newlib-nano-formatted-io,
floating-point support is split out of the formatted I/O code into
weak functions which are not linked by default. This leads to a smaller
code  when using newlib "printf" and/or "sscanf" but not using floating
point I/O.

Programs that need floating-point I/O support must explicitly request
linking of one or both of the floating-point functions:

     _printf_float or _scanf_float.

This can be done at link time using the -u option which can be passed
to either gcc or ld.
In addition, make some additional configuration options common for all
newlibs, not just IAMCU, to keep various newlib configuration similar.
These options can lead to a smaller code as well (if the code uses wide
 characters, fseek, malloc, ...):

    --disable-newlib-wide-orient
    --disable-newlib-fseek-optimization
    --enable-newlib-nano-malloc

Change-Id: I3d4425d43ccd28e735d45fe4397215d1ecfb16fb
Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>