xtensa: remove unneeded patches
The patches that replace bb[cs]i.l with bb[cs]i are not needed, because
it's the assembler job.
The patch that changes XCHAL_HAVE_BE is inconsistent (only applied to gcc
and not to binutils/gdb where it wouldn't work) and is not needed,
because this definition comes from the CPU configuration overlay.

Drop these patches and references to them.

Change-Id: I8d5f182a239553edc125741293b669b25a7a631a
Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>